+++ Tests/CDUntappdKitTests/Decoder/CDUntappdURLSessionDecoderTests.swift
@@ -23,6 +23,7 @@
 //  THE SOFTWARE.
 //
 
+import CDUntappdKitTesting
 import Foundation
 import Testing
 @testable import CDUntappdKit